+++ waspc/examples/todoApp/src/client/pages/auth/Login.js
@@ -2,7 +2,8 @@ import React from 'react'
 import { Link } from 'react-router-dom'
 
 import LoginForm from '@wasp/auth/forms/Login'
-// import { GoogleSignInButton } from '@wasp/auth/buttons/Google'
+// import { SignInButton as GoogleSignInButton } from '@wasp/auth/helpers/Google'
+// import { SignInButton as GitHubSignInButton } from '@wasp/auth/helpers/GitHub'
 
 const Login = () => {
   return (
@@ -15,6 +16,7 @@ const Login = () => {
 
       {/* <div>
         <GoogleSignInButton/>
+        <GitHubSignInButton/>
       </div> */}
     </>
   )