+++ waspc/examples/todoApp/src/server/auth/google.js
@@ -3,7 +3,7 @@ import { generateAvailableUsername } from '@wasp/core/auth.js'
 export function config() {
   console.log("Inside user-supplied Google config")
   return {
-    clientId: process.env['GOOGLE_CLIENT_ID'],
+    clientID: process.env['GOOGLE_CLIENT_ID'],
     clientSecret: process.env['GOOGLE_CLIENT_SECRET'],
     scope: ['profile']
   }