+++ waspc/src/Wasp/Generator/WebAppGenerator.hs
@@ -15,6 +15,7 @@ import StrongPath
     Posix,
     Rel,
     relDirToPosix,
+    reldir,
     relfile,
     (</>),
   )
@@ -35,7 +36,12 @@ import Wasp.Generator.Monad (Generator)
 import qualified Wasp.Generator.NpmDependencies as N
 import Wasp.Generator.WebAppGenerator.AuthG (genAuth)
 import qualified Wasp.Generator.WebAppGenerator.Common as C
-import Wasp.Generator.WebAppGenerator.ExternalCodeGenerator (extClientCodeDirInWebAppSrcDir, extClientCodeGeneratorStrategy, extSharedCodeGeneratorStrategy)
+import Wasp.Generator.WebAppGenerator.ExternalAuthG (ExternalAuthInfo (..), gitHubAuthInfo, googleAuthInfo)
+import Wasp.Generator.WebAppGenerator.ExternalCodeGenerator
+  ( extClientCodeDirInWebAppSrcDir,
+    extClientCodeGeneratorStrategy,
+    extSharedCodeGeneratorStrategy,
+  )
 import Wasp.Generator.WebAppGenerator.OperationsGenerator (genOperations)
 import Wasp.Generator.WebAppGenerator.RouterGenerator (genRouter)
 import Wasp.Util ((<++>))
@@ -147,23 +153,32 @@ genGitignore =
 genPublicDir :: AppSpec -> Generator [FileDraft]
 genPublicDir spec = do
   publicIndexHtmlFd <- genPublicIndexHtml spec
-  return $
+  return
     [ publicIndexHtmlFd,
       genFaviconFd,
       genManifestFd
     ]
-      ++ genGoogleSigninImage
+    <++> genSocialLoginIcons maybeAuth
   where
     maybeAuth = AS.App.auth $ snd $ getApp spec
     genFaviconFd = C.mkTmplFd (C.asTmplFile [relfile|public/favicon.ico|])
-    genGoogleSigninImage =
-      [ C.mkTmplFd (C.asTmplFile [relfile|public/images/btn_google_signin_dark_normal_web@2x.png|])
-        | (AS.App.Auth.isGoogleAuthEnabled <$> maybeAuth) == Just True
-      ]
     genManifestFd =
       let tmplData = object ["appName" .= (fst (getApp spec) :: String)]
           tmplFile = C.asTmplFile [relfile|public/manifest.json|]
        in C.mkTmplFdWithData tmplFile tmplData
+
+genSocialLoginIcons :: Maybe AS.App.Auth.Auth -> Generator [FileDraft]
+genSocialLoginIcons maybeAuth =
+  return $
+    [ C.mkTmplFd (C.asTmplFile fp)
+      | (isEnabled, fp) <- socialIcons,
+        (isEnabled <$> maybeAuth) == Just True
+    ]
+  where
+    socialIcons =
+      [ (AS.App.Auth.isGoogleAuthEnabled, [reldir|public/images|] </> _logoFileName googleAuthInfo),
+        (AS.App.Auth.isGitHubAuthEnabled, [reldir|public/images|] </> _logoFileName gitHubAuthInfo)
+      ]
 
 genPublicIndexHtml :: AppSpec -> Generator FileDraft
 genPublicIndexHtml spec =